+++ e2e-tests/endpoints/westend/accounts/balance-info/1191.json
@@ -1,0 +1,13 @@
+{
+    "at": {
+        "hash": "0x086577d296480f1808fda83a2a44fc4f436aef3bfd46b1fdfb48554cc2ba798d",
+        "height": "1191"
+    },
+    "nonce": "0",
+    "tokenSymbol": "WND",
+    "free": "50000000000000",
+    "reserved": "0",
+    "miscFrozen": "0",
+    "feeFrozen": "0",
+    "locks": []
+}
