+++ e2e-tests/endpoints/westend/accounts/balance-info/7245493.json
@@ -1,0 +1,13 @@
+{
+    "at": {
+        "hash": "0xaf7df2540519ff86aae7fe29dd9864f7600e22ba839505d26a830bd3e5c906f0",
+        "height": "7245493"
+    },
+    "nonce": "28023",
+    "tokenSymbol": "WND",
+    "free": "2963810663562321",
+    "reserved": "0",
+    "miscFrozen": "0",
+    "feeFrozen": "0",
+    "locks": []
+}
