+++ e2e-tests/endpoints/westend/accounts/index.ts
@@ -1,0 +1,3 @@
+import { westendAccountBalanceEndpoints } from './balance-info';
+
+export const westendAccountsEndpoints = [...westendAccountBalanceEndpoints];
